+++ lensless/recon/mirflickr.py
@@ -7,7 +7,7 @@
 
 
 import numpy as np
-from lensless.admm import ADMM
+from lensless.recon.admm import ADMM
 import cv2
