+++ profile/admm.py
@@ -1,5 +1,5 @@
 from diffusercam_admm import apply_admm
-from lensless.admm import ADMM
+from lensless.recon.admm import ADMM
 from lensless.io import load_data
 import time
 import pathlib as plib
