+++ scripts/evaluate_mirflickr_admm.py
@@ -17,10 +17,10 @@ import pathlib as plib
 from datetime import datetime
 from lensless.io import load_psf
 import numpy as np
-from lensless.util import print_image_info
+from lensless.image_utils import print_image_info
 from PIL import Image
-from lensless.mirflickr import ADMM_MIRFLICKR, postprocess
-from lensless.metric import mse, psnr, ssim, lpips
+from lensless.recon.mirflickr import ADMM_MIRFLICKR, postprocess
+from lensless.eval.metric import mse, psnr, ssim, lpips
 
 
 @hydra.main(version_base=None, config_path="../configs", config_name="evaluate_mirflickr_admm")
